Fixed ossie #255
